+++ src/lib/builders/context-menu/create.ts
@@ -48,7 +48,7 @@ const defaults = {
 	preventScroll: true,
 	escapeBehavior: 'close',
 	interactOutsideBehavior: 'close',
-	portal: undefined,
+	portal: 'body',
 	loop: false,
 	dir: 'ltr',
 	defaultOpen: false,